Buttons
This code allows the number to increase or decrease depending on which
button is pressed. It changes the color of the fan speed button(s) and
the text
when it is pressed/unpressed. It also turns the switches on/off. This
code goes under java Main Activity in
Android Studio